requirements.txt
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-google-genai>=2.0.0,<3.0.0
langchain-community>=0.3.0,<0.4.0
python-dotenv>=1.0.0
streamlit>=1.30.0
pandas>=2.0.0
pydantic>=2.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
